Moved command utilites to `module/utils/CommandUtils`
