--- modules/util/_Commands.py
@@ -1,18 +1,0 @@
-from commands import (
-    run, banner, clear, show,
-    doc, set, unset, use,
-    help, back
-)
-
-Commands: dict = {
-    'run': run.run,
-    'banner': banner.run,
-    'clear': clear.run,
-    'show': show.run,
-    'doc': doc.Docs,
-    'set': set.Set,
-    'unset': unset.unset_val,
-    'use': use.use,
-    'help': help.run,
-    'back': back.run
-}
